AI Girlfriend/requirements.txt
python-dotenv
sounddevice
scipy
numpy
elevenlabs==1.5.0
langchain
langchain-groq